+++ doc/code/utils.rst
@@ -1,0 +1,15 @@
+mrmustard.utils
+===============
+
+.. toctree::
+    :maxdepth: 1
+
+    utils/settings
+    utils/typing
+
+
+.. currentmodule:: mrmustard.utils
+
+.. automodapi:: mrmustard.utils
+    :no-heading:
+    :include-all-objects:
