+++ doc/code/utils/settings.rst
@@ -1,0 +1,8 @@
+Settings
+========
+
+.. currentmodule:: mrmustard.utils.settings
+
+.. automodapi:: mrmustard.utils.settings
+    :no-heading:
+    :include-all-objects:
