+++ doc/code/utils/typing.rst
@@ -1,0 +1,8 @@
+Typing
+======
+
+.. currentmodule:: mrmustard.utils.typing
+
+.. automodapi:: mrmustard.utils.typing
+    :no-heading:
+    :include-all-objects:
