+++ doc/index.rst
@@ -131,3 +131,4 @@ Mr Mustard supports the following in a fully differentiable way:
    code/physics
    code/math
    code/training
+   code/utils
