+++ mrmustard/lab/abstract/transformation.py
@@ -28,7 +28,7 @@ from mrmustard.math import Math
 from mrmustard.math.tensor_networks import Tensor
 from mrmustard.physics import bargmann, fock, gaussian
 from mrmustard.training.parameter import Parameter
-from mrmustard.typing import RealMatrix, RealVector
+from mrmustard.utils.typing import RealMatrix, RealVector
 
 from .state import State
