+++ mrmustard/math/lattice/steps.py
@@ -28,7 +28,7 @@ from numba.cpython.unsafe.tuple import tuple_setitem
 
 from mrmustard.math.lattice.neighbors import lower_neighbors
 from mrmustard.math.lattice.pivots import first_available_pivot
-from mrmustard.typing import ComplexMatrix, ComplexTensor, ComplexVector
+from mrmustard.utils.typing import ComplexMatrix, ComplexTensor, ComplexVector
 
 SQRT = np.sqrt(np.arange(100000))
