+++ mrmustard/math/tensorflow_wrapper.py
@@ -29,8 +29,8 @@ from mrmustard.math.compactFock.compactFock_inputValidation import (
     hermite_multidimensional_1leftoverMode,
     hermite_multidimensional_diagonal,
 )
-from mrmustard.typing import Tensor, Trainable
-from mrmustard.math.math_interface import MathInterface
+from mrmustard.utils.typing import Tensor, Trainable
+from .math_interface import MathInterface
 
 
 # pylint: disable=too-many-public-methods,no-self-argument,arguments-differ
