+++ mrmustard/training/parameter_update.py
@@ -17,7 +17,7 @@
 
 from typing import Tuple, Sequence
 from mrmustard.math import Math
-from mrmustard.typing import Tensor
+from mrmustard.utils.typing import Tensor
 
 from .parameter import Trainable
