+++ mrmustard/utils/settings.py
@@ -297,8 +297,9 @@ class Settings:
             from julia import Main as Main_julia  # pylint: disable=import-outside-toplevel
 
             # import Julia functions
-            math_directory = os.path.dirname(__file__)
-            Main_julia.cd(math_directory)
+            utils_directory = os.path.dirname(__file__)
+            mm_directory = os.path.dirname(utils_directory)
+            Main_julia.cd(mm_directory)
             Main_julia.include("math/lattice/strategies/vanilla.jl")
 
     # use rich.table to print the settings
