+++ mrmustard/utils/typing.py
