+++ tests/test_math/test_xptensor.py
@@ -12,13 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from hypothesis import strategies as st, given, assume
+from hypothesis import strategies as st, given
 from hypothesis.extra.numpy import arrays
-import pytest
-from mrmustard.lab.states import DisplacedSqueezed
-from mrmustard.utils.xptensor import XPVector, XPMatrix
+from mrmustard.math.xptensor import XPVector, XPMatrix
 import numpy as np
-from tests.random import n_mode_pure_state
 
 even = st.integers(min_value=2, max_value=10).filter(lambda x: x % 2 == 0)
 floats = st.floats(min_value=-1e3, max_value=1e3, allow_nan=False, allow_infinity=False)
