+++ tests/test_utils/test_logger.py
@@ -50,7 +50,7 @@ https://github.com/pallets/flask/blob/master/tests/test_logging.py
 import logging
 import pytest
 from mrmustard.training import optimizer
-from mrmustard.logger import logging_handler_defined, default_handler, create_logger
+from mrmustard.utils.logger import logging_handler_defined, default_handler, create_logger
 
 modules_contain_logging = [optimizer]
